requirements.txt
streamlit==1.32.0
requests==2.31.0
beautifulsoup4==4.12.3

sentence-transformers==2.2.2
huggingface_hub==0.19.4

faiss-cpu==1.7.4

groq==0.5.0
python-dotenv==1.0.1
pickle-mixin==1.0.2
numpy==1.26.4
torch==2.1.2
